ETT: Make GetRepresentative return a vertex (#10)
Make `ETT.GetRepresentative()` return a vertex ID rather than an arbitrary skip list element. It may be a bit slower to retrieve the vertex ID, but a vertex ID is easier to use (e.g., easy to place in a hash table, e.g., if we link two components with representative vertices a and b, the linked components will have representative vertex a or b).