ETT: Re-add function for getting vertices of component
In PR #9 I removed the old function for getting vertices of a component
and replaced it with the function ComponentEdges that gets edges of a
component. In the one place where we do look at the vertices of a
component, I changed it to just use ComponentEdges. But there's a bug
there when ComponentEdges is empty, so in this commit I'm adding a
naive, sequential implementation of ComponentVertices using
ComponentEdges that handles that edge case.