augmented_skip_list: Avoid segfault by increasing size of augmented value
Augmented skip lists were failing because AugmentedElement::values_
would sometimes be 4-bytes large, and that would break block_allocator
(see https://github.com/cmuparlay/parlaylib/pull/8).

This commit makes makes `values_` bigger in a naive way: by making the
augmented value a `long long` rather than an `int`. We shouldn't bother
doing anything smarter since we're going to replace the augmented value
with a more general template in the future anyway. I decided to do some
of that templatization in this commit as well to get a head start on it.